test: redis connection and functions

+++ internal/timeline/timeline.go
@@ -84,7 +84,7 @@ func (tl *TimelineService) fanout(ctx context.Context, userID, postID uint, scor
 }
 
 func (tl *TimelineService) GetTimeLine(ctx context.Context, userId, limit, offset uint) ([]dataaccess.Post, error) {
-	postIds, err := tl.cache.GetTimelinePostIDs(ctx, userId, int64(offset), int64(offset)+int64(limit))
+	postIds, err := tl.cache.GetTimelinePostIDs(ctx, userId, offset, offset+limit)
 	if err != nil {
 		err = fmt.Errorf("failed to get user=%d timeline:%w", userId, err)
 		return nil, err
